+++ lib/Sema/CSDiag.h
@@ -19,8 +19,6 @@
 #define SWIFT_SEMA_CSDIAG_H
 
 #include "ConstraintSystem.h"
-#include "swift/AST/DiagnosticEngine.h"
-#include "llvm/ADT/StringRef.h"
 #include <string>
 
 namespace swift {
@@ -34,12 +32,6 @@ namespace swift {
   /// UnresolvedType.
   Type replaceTypeParametersWithUnresolved(Type ty);
   Type replaceTypeVariablesWithUnresolved(Type ty);
-
-  /// Diagnose lvalue expr error.
-  void diagnoseSubElementFailure(Expr *destExpr, SourceLoc loc,
-                                 constraints::ConstraintSystem &CS,
-                                 Diag<StringRef> diagID,
-                                 Diag<Type> unknownDiagID);
 };
 
 #endif /* SWIFT_SEMA_CSDIAG_H */
